More css enhancements.
This latest check-in allows users to reference images in the css files as they normally would and the CssManager will automatically resolve the url.

For example: url (../images/imagename.jpg) will now automatically resolve to the correct location within the blob storage. This is a significant enhancement, because it ensure that image references within css files work and work as users would expect them to

i.e. no need to worry about blob storage, or where the image is actually stored, they just use ../images/<filename> and their images will automatically be resolved for them