Modified preview to not actually save the page.
Uses a queue to transfer the data and then
pulls the page from cache and updates the content.

It now takes ~4 seconds to load the preview, down from 8 seconds. There are a few other tweaks I can make to speed it up further; the delay now seems to be caused by the database (which makes sense since I'm hitting the azure database and not a local one)