More work on the site/theme store.
Added ability to attach screenshots to the package.
Integrating the actual store front into the backend.

There is about 6-7hrs of work left, which I will be finishing up tomorrow afternoon and evening. This will leave us Tuesday and Wednesday to test, and deploy to Azure for the VC demo.