Fix an issue in the MS Azure Storage Client which caused the container to show as "not available"
This is a bug in the Microsoft Azure Storage API. To work around it, I wrote an extension method which checks if the
container exists and if not, it tries to create it. It then handles the exception which can occur if you had immeidately deleted
the container prior to trying to re-create it. In this case, the API would report "true" as the container being created, when
in fact, it had failed.