Removed the ajaxification of this page (for now).
For some reason the ASP.NET AJAX update panel and Google Chrome do not get along together. I have not yet figured out a workaround. I am going to write a simple proof-of-concept and open a telerik support ticket. In the meantime, removing the ajax update panel seems to fix the problem.

bugzid: 168