I finally was able to determine where the conflict was at... it's with mootools-core.js
I do not believe we use this set of javascript functions any longer... but maybe we do? If we do, we need to find them and modify them to use dojo or jquery instead.

I have removed mootools from the master page, so keep an eye out for any javascript errors or things that just "don't work"

bugzid: 168